# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChapterProgress {
  id: number;
  chapterId: number;
  completed: boolean;
  sectionsCompleted: string[];
  /** @nullable */
  lastVisited: string | null;
  /** @nullable */
  quizBestScore: number | null;
  updatedAt: string;
}

export interface ProgressSummary {
  totalChapters: number;
  completedChapters: number;
  percentComplete: number;
  chaptersProgress: ChapterProgress[];
}

export interface ProgressUpdate {
  completed?: boolean;
  sectionsCompleted?: string[];
  lastVisited?: string;
}

export type QuizQuestionType = typeof QuizQuestionType[keyof typeof QuizQuestionType];


export const QuizQuestionType = {
  multiple_choice: 'multiple_choice',
  short_answer: 'short_answer',
} as const;

export interface QuizQuestion {
  id: number;
  chapterId: number;
  type: QuizQuestionType;
  question: string;
  /** @nullable */
  options: string[] | null;
  correctAnswer: string;
  explanation: string;
  order: number;
}

export interface QuizAnswer {
  questionId: number;
  answer: string;
}

export interface QuizSubmission {
  answers: QuizAnswer[];
}

export interface QuestionFeedback {
  questionId: number;
  question: string;
  userAnswer: string;
  correctAnswer: string;
  isCorrect: boolean;
  explanation: string;
}

export interface QuizResult {
  id: number;
  chapterId: number;
  score: number;
  total: number;
  percentage: number;
  feedback: QuestionFeedback[];
  submittedAt: string;
}

export interface Bookmark {
  id: number;
  chapterId: number;
  /** @nullable */
  sectionId: string | null;
  title: string;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface BookmarkInput {
  chapterId: number;
  /** @nullable */
  sectionId?: string | null;
  title: string;
  /** @nullable */
  note?: string | null;
}

export interface GlossaryTerm {
  id: number;
  term: string;
  definition: string;
  /** @nullable */
  chapterId: number | null;
  /** @nullable */
  category: string | null;
}

export type SearchResultType = typeof SearchResultType[keyof typeof SearchResultType];


export const SearchResultType = {
  topic: 'topic',
  definition: 'definition',
  formula: 'formula',
  glossary: 'glossary',
} as const;

export interface SearchResult {
  type: SearchResultType;
  title: string;
  excerpt: string;
  /** @nullable */
  chapterId: number | null;
  /** @nullable */
  chapterTitle: string | null;
  /** @nullable */
  sectionId: string | null;
}

export interface SearchResults {
  query: string;
  results: SearchResult[];
}

export type GetGlossaryTermsParams = {
/**
 * @nullable
 */
chapterId?: number | null;
};

export type SearchContentParams = {
q: string;
};