# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PetroLearn API - Reservoir Petrophysics Educational Platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: progress
    description: Chapter progress tracking
  - name: quiz
    description: Chapter quizzes and results
  - name: bookmarks
    description: User bookmarks
  - name: glossary
    description: Glossary terms
  - name: search
    description: Global search

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /progress:
    get:
      operationId: getAllProgress
      tags: [progress]
      summary: Get progress for all chapters
      responses:
        "200":
          description: All chapter progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChapterProgress"

  /progress/summary:
    get:
      operationId: getProgressSummary
      tags: [progress]
      summary: Get overall course progress summary
      responses:
        "200":
          description: Overall progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressSummary"

  /progress/{chapterId}:
    get:
      operationId: getChapterProgress
      tags: [progress]
      summary: Get progress for a specific chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterProgress"
        "404":
          description: Not found

    patch:
      operationId: updateChapterProgress
      tags: [progress]
      summary: Update chapter progress
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterProgress"

  /quiz/{chapterId}:
    get:
      operationId: getChapterQuiz
      tags: [quiz]
      summary: Get quiz questions for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"
        "404":
          description: Not found

  /quiz/{chapterId}/submit:
    post:
      operationId: submitQuiz
      tags: [quiz]
      summary: Submit quiz answers and get results
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result with feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  /quiz/{chapterId}/history:
    get:
      operationId: getQuizHistory
      tags: [quiz]
      summary: Get past quiz results for a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizResult"

  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: Get all user bookmarks
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"

    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"

  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /glossary:
    get:
      operationId: getGlossaryTerms
      tags: [glossary]
      summary: Get all glossary terms
      parameters:
        - name: chapterId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of glossary terms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GlossaryTerm"

  /search:
    get:
      operationId: searchContent
      tags: [search]
      summary: Search across topics, definitions, and formulas
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ChapterProgress:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        completed:
          type: boolean
        sectionsCompleted:
          type: array
          items:
            type: string
        lastVisited:
          type: ["string", "null"]
        quizBestScore:
          type: ["integer", "null"]
        updatedAt:
          type: string
      required:
        - id
        - chapterId
        - completed
        - sectionsCompleted
        - lastVisited
        - quizBestScore
        - updatedAt

    ProgressSummary:
      type: object
      properties:
        totalChapters:
          type: integer
        completedChapters:
          type: integer
        percentComplete:
          type: integer
        chaptersProgress:
          type: array
          items:
            $ref: "#/components/schemas/ChapterProgress"
      required:
        - totalChapters
        - completedChapters
        - percentComplete
        - chaptersProgress

    ProgressUpdate:
      type: object
      properties:
        completed:
          type: boolean
        sectionsCompleted:
          type: array
          items:
            type: string
        lastVisited:
          type: string

    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        type:
          type: string
          enum: [multiple_choice, short_answer]
        question:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        order:
          type: integer
      required:
        - id
        - chapterId
        - type
        - question
        - options
        - correctAnswer
        - explanation
        - order

    QuizSubmission:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
      required:
        - answers

    QuizAnswer:
      type: object
      properties:
        questionId:
          type: integer
        answer:
          type: string
      required:
        - questionId
        - answer

    QuizResult:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        score:
          type: integer
        total:
          type: integer
        percentage:
          type: integer
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuestionFeedback"
        submittedAt:
          type: string
      required:
        - id
        - chapterId
        - score
        - total
        - percentage
        - feedback
        - submittedAt

    QuestionFeedback:
      type: object
      properties:
        questionId:
          type: integer
        question:
          type: string
        userAnswer:
          type: string
        correctAnswer:
          type: string
        isCorrect:
          type: boolean
        explanation:
          type: string
      required:
        - questionId
        - question
        - userAnswer
        - correctAnswer
        - isCorrect
        - explanation

    Bookmark:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        sectionId:
          type: ["string", "null"]
        title:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - chapterId
        - sectionId
        - title
        - note
        - createdAt

    BookmarkInput:
      type: object
      properties:
        chapterId:
          type: integer
        sectionId:
          type: ["string", "null"]
        title:
          type: string
        note:
          type: ["string", "null"]
      required:
        - chapterId
        - title

    GlossaryTerm:
      type: object
      properties:
        id:
          type: integer
        term:
          type: string
        definition:
          type: string
        chapterId:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
      required:
        - id
        - term
        - definition
        - chapterId
        - category

    SearchResults:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
      required:
        - query
        - results

    SearchResult:
      type: object
      properties:
        type:
          type: string
          enum: [topic, definition, formula, glossary]
        title:
          type: string
        excerpt:
          type: string
        chapterId:
          type: ["integer", "null"]
        chapterTitle:
          type: ["string", "null"]
        sectionId:
          type: ["string", "null"]
      required:
        - type
        - title
        - excerpt
        - chapterId
        - chapterTitle
        - sectionId