# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  chapterId: number;
  /** @nullable */
  sectionId: string | null;
  title: string;
  /** @nullable */
  note: string | null;
  createdAt: string;
}