# lib/api-zod/src/generated/types/chapterProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */

export interface ChapterProgress {
  id: number;
  chapterId: number;
  completed: boolean;
  sectionsCompleted: string[];
  /** @nullable */
  lastVisited: string | null;
  /** @nullable */
  quizBestScore: number | null;
  updatedAt: string;
}