# lib/api-zod/src/generated/types/glossaryTerm.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */

export interface GlossaryTerm {
  id: number;
  term: string;
  definition: string;
  /** @nullable */
  chapterId: number | null;
  /** @nullable */
  category: string | null;
}