# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterProgress } from './chapterProgress';

export interface ProgressSummary {
  totalChapters: number;
  completedChapters: number;
  percentComplete: number;
  chaptersProgress: ChapterProgress[];
}