# lib/api-zod/src/generated/types/progressUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressUpdate {
  completed?: boolean;
  sectionsCompleted?: string[];
  lastVisited?: string;
}