# lib/api-zod/src/generated/types/questionFeedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionFeedback {
  questionId: number;
  question: string;
  userAnswer: string;
  correctAnswer: string;
  isCorrect: boolean;
  explanation: string;
}