# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResultType } from './searchResultType';

export interface SearchResult {
  type: SearchResultType;
  title: string;
  excerpt: string;
  /** @nullable */
  chapterId: number | null;
  /** @nullable */
  chapterTitle: string | null;
  /** @nullable */
  sectionId: string | null;
}