# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetroLearn API - Reservoir Petrophysics Educational Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from './searchResult';

export interface SearchResults {
  query: string;
  results: SearchResult[];
}